Rename TripArrivalVieww to ArrivalDepartureView
